+++ client-flink/common/src/main/java/org/apache/celeborn/plugin/flink/RemoteShuffleEnvironment.java
@@ -20,12 +20,15 @@ package org.apache.celeborn.plugin.flink;
 import static org.apache.celeborn.plugin.flink.metric.RemoteShuffleMetricFactory.createShuffleIOOwnerMetricGroup;
 import static org.apache.celeborn.plugin.flink.utils.Utils.checkNotNull;
 import static org.apache.celeborn.plugin.flink.utils.Utils.checkState;
-import static org.apache.flink.runtime.io.network.metrics.NettyShuffleMetricFactory.*;
+import static org.apache.flink.runtime.io.network.metrics.NettyShuffleMetricFactory.METRIC_GROUP_INPUT;
+import static org.apache.flink.runtime.io.network.metrics.NettyShuffleMetricFactory.METRIC_GROUP_OUTPUT;
 
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.stream.Collectors;
 
 import org.apache.flink.annotation.VisibleForTesting;
 import org.apache.flink.metrics.MetricGroup;
@@ -35,10 +38,13 @@ import org.apache.flink.runtime.executiongraph.ExecutionAttemptID;
 import org.apache.flink.runtime.executiongraph.PartitionInfo;
 import org.apache.flink.runtime.io.network.api.writer.ResultPartitionWriter;
 import org.apache.flink.runtime.io.network.buffer.NetworkBufferPool;
+import org.apache.flink.runtime.io.network.metrics.InputChannelMetrics;
 import org.apache.flink.runtime.io.network.partition.PartitionProducerStateProvider;
 import org.apache.flink.runtime.io.network.partition.ResultPartitionID;
 import org.apache.flink.runtime.io.network.partition.ResultPartitionManager;
 import org.apache.flink.runtime.io.network.partition.consumer.IndexedInputGate;
+import org.apache.flink.runtime.jobgraph.IntermediateDataSetID;
+import org.apache.flink.runtime.jobgraph.IntermediateResultPartitionID;
 import org.apache.flink.runtime.shuffle.ShuffleEnvironment;
 import org.apache.flink.runtime.shuffle.ShuffleIOOwnerContext;
 import org.apache.flink.util.FlinkRuntimeException;
@@ -46,43 +52,64 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import org.apache.celeborn.common.CelebornConf;
+import org.apache.celeborn.plugin.flink.netty.NettyShuffleEnvironmentWrapper;
 
 /**
  * The implementation of {@link ShuffleEnvironment} based on the remote shuffle service, providing
  * shuffle environment on flink TM side.
  */
-public abstract class AbstractRemoteShuffleEnvironment {
+public class RemoteShuffleEnvironment
+    implements ShuffleEnvironment<ResultPartitionWriter, IndexedInputGate> {
 
-  private static final Logger LOG = LoggerFactory.getLogger(AbstractRemoteShuffleEnvironment.class);
+  private static final Logger LOG = LoggerFactory.getLogger(RemoteShuffleEnvironment.class);
 
   /** Network buffer pool for shuffle read and shuffle write. */
-  protected final NetworkBufferPool networkBufferPool;
+  private final NetworkBufferPool networkBufferPool;
 
   /** A trivial {@link ResultPartitionManager}. */
-  protected final ResultPartitionManager resultPartitionManager;
+  private final ResultPartitionManager resultPartitionManager;
+
+  /** Factory class to create {@link AbstractRemoteShuffleResultPartition}. */
+  private final AbstractRemoteShuffleResultPartitionFactory resultPartitionFactory;
+
+  private final AbstractRemoteShuffleInputGateFactory inputGateFactory;
+
+  private final CelebornConf conf;
+
+  private final NettyShuffleEnvironmentWrapper shuffleEnvironmentWrapper;
 
   /** Whether the shuffle environment is closed. */
-  protected boolean isClosed;
+  private boolean isClosed;
 
-  protected final Object lock = new Object();
+  private final Object lock = new Object();
 
-  protected final CelebornConf conf;
+  private final ConcurrentHashMap.KeySetView<IntermediateDataSetID, Boolean> nettyResultIds =
+      ConcurrentHashMap.newKeySet();
+
+  private final ConcurrentHashMap.KeySetView<IntermediateResultPartitionID, Boolean>
+      nettyResultPartitionIds = ConcurrentHashMap.newKeySet();
 
   /**
    * @param networkBufferPool Network buffer pool for shuffle read and shuffle write.
    * @param resultPartitionManager A trivial {@link ResultPartitionManager}.
    */
-  public AbstractRemoteShuffleEnvironment(
+  public RemoteShuffleEnvironment(
       NetworkBufferPool networkBufferPool,
       ResultPartitionManager resultPartitionManager,
-      CelebornConf conf) {
-
+      AbstractRemoteShuffleResultPartitionFactory resultPartitionFactory,
+      AbstractRemoteShuffleInputGateFactory inputGateFactory,
+      CelebornConf conf,
+      NettyShuffleEnvironmentWrapper shuffleEnvironmentWrapper) {
     this.networkBufferPool = networkBufferPool;
     this.resultPartitionManager = resultPartitionManager;
+    this.resultPartitionFactory = resultPartitionFactory;
+    this.inputGateFactory = inputGateFactory;
     this.conf = conf;
+    this.shuffleEnvironmentWrapper = shuffleEnvironmentWrapper;
     this.isClosed = false;
   }
 
+  @Override
   public void close() {
     LOG.info("Close RemoteShuffleEnvironment.");
     synchronized (lock) {
@@ -105,6 +132,7 @@ public abstract class AbstractRemoteShuffleEnvironment {
     }
   }
 
+  @Override
   public int start() throws IOException {
     synchronized (lock) {
       checkState(!isClosed, "The RemoteShuffleEnvironment has already been shut down.");
@@ -114,10 +142,12 @@ public abstract class AbstractRemoteShuffleEnvironment {
     }
   }
 
+  @Override
   public boolean updatePartitionInfo(ExecutionAttemptID consumerID, PartitionInfo partitionInfo) {
     throw new FlinkRuntimeException("Not implemented yet.");
   }
 
+  @Override
   public ShuffleIOOwnerContext createShuffleIOOwnerContext(
       String ownerName, ExecutionAttemptID executionAttemptID, MetricGroup parentGroup) {
     MetricGroup remoteGroup = createShuffleIOOwnerMetricGroup(checkNotNull(parentGroup));
@@ -129,15 +159,12 @@ public abstract class AbstractRemoteShuffleEnvironment {
         remoteGroup.addGroup(METRIC_GROUP_INPUT));
   }
 
+  @Override
   public Collection<ResultPartitionID> getPartitionsOccupyingLocalResources() {
     return resultPartitionManager.getUnreleasedPartitions();
   }
 
-  @VisibleForTesting
-  NetworkBufferPool getNetworkBufferPool() {
-    return networkBufferPool;
-  }
-
+  @Override
   public List<ResultPartitionWriter> createResultPartitionWriters(
       ShuffleIOOwnerContext ownerContext,
       List<ResultPartitionDeploymentDescriptor> resultPartitionDeploymentDescriptors) {
@@ -156,12 +183,38 @@ public abstract class AbstractRemoteShuffleEnvironment {
     }
   }
 
-  public abstract ResultPartitionWriter createResultPartitionWriterInternal(
+  private ResultPartitionWriter createResultPartitionWriterInternal(
       ShuffleIOOwnerContext ownerContext,
       int index,
       ResultPartitionDeploymentDescriptor resultPartitionDeploymentDescriptor,
-      CelebornConf conf);
+      CelebornConf conf) {
+    if (resultPartitionDeploymentDescriptor.getShuffleDescriptor()
+        instanceof RemoteShuffleDescriptor) {
+      return resultPartitionFactory.create(
+          ownerContext.getOwnerName(), index, resultPartitionDeploymentDescriptor, conf);
+    } else {
+      nettyResultIds.add(resultPartitionDeploymentDescriptor.getResultId());
+      nettyResultPartitionIds.add(resultPartitionDeploymentDescriptor.getPartitionId());
+      return shuffleEnvironmentWrapper
+          .nettyResultPartitionFactory()
+          .create(ownerContext.getOwnerName(), index, resultPartitionDeploymentDescriptor);
+    }
+  }
 
+  @Override
+  public void releasePartitionsLocally(Collection<ResultPartitionID> partitionIds) {
+    List<ResultPartitionID> resultPartitionIds =
+        partitionIds.stream()
+            .filter(partitionId -> nettyResultPartitionIds.contains(partitionId.getPartitionId()))
+            .collect(Collectors.toList());
+    if (!resultPartitionIds.isEmpty()) {
+      shuffleEnvironmentWrapper
+          .nettyShuffleEnvironment()
+          .releasePartitionsLocally(resultPartitionIds);
+    }
+  }
+
+  @Override
   public List<IndexedInputGate> createInputGates(
       ShuffleIOOwnerContext ownerContext,
       PartitionProducerStateProvider producerStateProvider,
@@ -180,9 +233,26 @@ public abstract class AbstractRemoteShuffleEnvironment {
     }
   }
 
-  abstract IndexedInputGate createInputGateInternal(
+  private IndexedInputGate createInputGateInternal(
       ShuffleIOOwnerContext ownerContext,
       PartitionProducerStateProvider producerStateProvider,
       int gateIndex,
-      InputGateDeploymentDescriptor igdd);
+      InputGateDeploymentDescriptor igdd) {
+    return nettyResultIds.contains(igdd.getConsumedResultId())
+        ? shuffleEnvironmentWrapper
+            .nettyInputGateFactory()
+            .create(
+                ownerContext,
+                gateIndex,
+                igdd,
+                producerStateProvider,
+                new InputChannelMetrics(
+                    ownerContext.getInputGroup(), ownerContext.getParentGroup()))
+        : inputGateFactory.create(ownerContext.getOwnerName(), gateIndex, igdd);
+  }
+
+  @VisibleForTesting
+  AbstractRemoteShuffleResultPartitionFactory getResultPartitionFactory() {
+    return resultPartitionFactory;
+  }
 }
