+++ client-flink/flink-1.16/src/test/java/org/apache/celeborn/plugin/flink/RemoteShuffleMasterSuiteJ.java
@@ -321,7 +321,6 @@ public class RemoteShuffleMasterSuiteJ {
                 System.exit(-1);
               }
             },
-            new SimpleResultPartitionAdapter(),
             nettyShuffleServiceFactory);
 
     return remoteShuffleMaster;
