+++ client-flink/flink-1.17/src/main/java/org/apache/celeborn/plugin/flink/RemoteShuffleServiceFactory.java
@@ -17,30 +17,14 @@
 
 package org.apache.celeborn.plugin.flink;
 
-import org.apache.flink.runtime.io.network.NettyShuffleServiceFactory;
 import org.apache.flink.runtime.io.network.api.writer.ResultPartitionWriter;
 import org.apache.flink.runtime.io.network.partition.consumer.IndexedInputGate;
-import org.apache.flink.runtime.shuffle.ShuffleDescriptor;
 import org.apache.flink.runtime.shuffle.ShuffleEnvironment;
 import org.apache.flink.runtime.shuffle.ShuffleEnvironmentContext;
-import org.apache.flink.runtime.shuffle.ShuffleMaster;
-import org.apache.flink.runtime.shuffle.ShuffleMasterContext;
-import org.apache.flink.runtime.shuffle.ShuffleServiceFactory;
 
 import org.apache.celeborn.plugin.flink.netty.NettyShuffleEnvironmentWrapper;
 
-public class RemoteShuffleServiceFactory extends AbstractRemoteShuffleServiceFactory
-    implements ShuffleServiceFactory<ShuffleDescriptor, ResultPartitionWriter, IndexedInputGate> {
-
-  private final NettyShuffleServiceFactory nettyShuffleServiceFactory =
-      new NettyShuffleServiceFactory();
-
-  @Override
-  public ShuffleMaster<ShuffleDescriptor> createShuffleMaster(
-      ShuffleMasterContext shuffleMasterContext) {
-    return new RemoteShuffleMaster(
-        shuffleMasterContext, new SimpleResultPartitionAdapter(), nettyShuffleServiceFactory);
-  }
+public class RemoteShuffleServiceFactory extends AbstractRemoteShuffleServiceFactory {
 
   @Override
   public ShuffleEnvironment<ResultPartitionWriter, IndexedInputGate> createShuffleEnvironment(
