+++ client-flink/flink-1.17/src/test/java/org/apache/celeborn/plugin/flink/RemoteShuffleMasterSuiteJ.java
@@ -333,7 +333,6 @@ public class RemoteShuffleMasterSuiteJ {
                 System.exit(-1);
               }
             },
-            new SimpleResultPartitionAdapter(),
             nettyShuffleServiceFactory);
 
     return remoteShuffleMaster;
